Refactor cli command, options parsing to a separate object
