+++ pakyow-core/lib/pakyow/command.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Pakyow
+  class Command
+  end
+end
